fix: 메인 페이지 에러 수정 (#29)
