+++ src/components/molecules/MainImg/MainImg.css
@@ -1,0 +1,46 @@
+.main_img {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  align-items: center;
+  /* background-image: url('https://www.gist.ac.kr/site/kr/img/main/mvisual_bg04.jpg'); */
+  background-image: url('assets/images/Gist.jpg');
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-attachment: scroll;
+  width: 100%;
+  height: 480px;
+
+  border-top: 1px solid #ccc;
+  border-bottom: 1px solid #ccc;
+}
+
+.main_img .main_slogan {
+  font-size: 34px;
+  /* font-weight: bold; */
+  text-align: center;
+  line-height: 48px;
+}
+
+.main_img .main_petition {
+  width: 240px;
+  height: 34px;
+}
+
+.main_img .main_petition a {
+  background-color: #df3127;
+  display: block;
+  width: 100%;
+  height: 100%;
+  font-size: 20px;
+  font-weight: bold;
+  line-height: 34px;
+  text-align: center;
+  box-sizing: border-box;
+
+  margin-top: 30px;
+
+  color: #fff;
+  text-decoration: none;
+}
