+++ src/components/molecules/MainNumOfPosts/MainNumOfPosts.css
@@ -1,0 +1,16 @@
+.main_num_of_posts {
+    margin: 80px 0 80px;
+  
+    text-align: center;
+    font-size: 34px;
+    font-weight: bold;
+    line-height: 48px;
+  }
+  
+  .main_num_of_posts span {
+    color: #df3127;
+    font-weight: bold;
+
+    
+
+  }
