--- src/pages/Main/MainNumOfPosts.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import './MainNumOfPosts.css';
-
-const MainNumOfPosts = () => {
-    return (
-      <div class="main_num_of_posts">
-        지금 까지<br/>총 <span>012</span>건의 청원과 <span>04</span>건의 답변이<br/>올라왔습니다
-      </div>
-    );
-    
-}
-
-
-export default MainNumOfPosts;
